requirements.txt
tensorflow~=2.20.0
numpy~=2.3.5
fastapi~=0.122.0
pillow~=12.0.0
gTTS~=2.5.4
groq~=0.36.0
dotenv~=0.9.9
python-dotenv~=1.2.1
requests~=2.32.5
mcp~=1.22.0
streamlit~=1.51.0